#29 認証部分をconcernにまとめた
